+++ phenonaut/transforms/__init__.py
@@ -6,3 +6,4 @@ from .transformer import *
 from .supervised_transformer import *
 from .dimensionality_reduction import PCA, TSNE, UMAP
 from .preparative import *
+from .imputers import *